LPS-203260 Improve packaging of the interfaces/classes in the package com.liferay.site.util
# breaking

## What modules/apps/site/site-api/src/main/java/com/liferay/site/util/RecentGroupManager.java

modules/apps/site/site-api/src/main/java/com/liferay/site/util/RecentGroupManager.java was moved from com.liferay.site.util package to the com.liferay.site.manager package

## Why

This change was made so the package matches the class name

----

# breaking

## What modules/apps/site/site-api/src/main/java/com/liferay/site/util/GroupSearchProvider.java

modules/apps/site/site-api/src/main/java/com/liferay/site/util/GroupSearchProvider.java was moved from com.liferay.site.util package to the com.liferay.site.provider package

## Why

This change was made so the package matches the class name

----

# breaking

## What modules/apps/site/site-api/src/main/java/com/liferay/site/util/GroupURLProvider.java

modules/apps/site/site-api/src/main/java/com/liferay/site/util/GroupURLProvider.java was moved from com.liferay.site.util package to the com.liferay.site.provider package

## Why

This change was made so the package matches the class name

----

# breaking

## What modules/apps/site/site-api/src/main/java/com/liferay/site/util/SitemapURLProvider.java

modules/apps/site/site-api/src/main/java/com/liferay/site/util/SitemapURLProvider.java was moved from com.liferay.site.util package to the com.liferay.site.provider package

## Why

This change was made so the package matches the class name

----

# breaking

## What modules/apps/site/site-api/src/main/java/com/liferay/site/util/SitemapURLProviderHelper.java

modules/apps/site/site-api/src/main/java/com/liferay/site/util/SitemapURLProviderHelper.java was moved from com.liferay.site.util package to the com.liferay.site.provider.helper package

## Why

This change was made so the package matches the class name

----
