+++ modules/apps/depot/depot-web/src/main/java/com/liferay/depot/web/internal/servlet/taglib/DepotBreadcrumbEntryContributorImpl.java
@@ -24,7 +24,7 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.site.util.GroupURLProvider;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.util.ArrayList;
 import java.util.List;
