+++ modules/apps/product-navigation/product-navigation-applications-menu-web-test/src/testIntegration/java/com/liferay/product/navigation/applications/menu/web/internal/portlet/action/test/ApplicationsMenuPanelAppsMVCResourceCommandTest.java
@@ -45,7 +45,7 @@ import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 import com.liferay.portletmvc4spring.test.mock.web.portlet.MockPortletRequest;
 import com.liferay.product.navigation.applications.menu.web.internal.portlet.action.test.constants.ApplicationsMenuTestPortletKeys;
-import com.liferay.site.util.RecentGroupManager;
+import com.liferay.site.manager.RecentGroupManager;
 
 import java.util.ArrayList;
 import java.util.List;
