+++ modules/apps/site/site-api/src/main/java/com/liferay/site/provider/helper/SitemapURLProviderHelper.java
@@ -1,0 +1,20 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.site.provider.helper;
+
+import com.liferay.portal.kernel.model.Layout;
+
+import org.osgi.annotation.versioning.ProviderType;
+
+/**
+ * @author Jürgen Kappler
+ */
+@ProviderType
+public interface SitemapURLProviderHelper {
+
+	public boolean isExcludeLayoutFromSitemap(Layout layout);
+
+}
