--- modules/apps/site/site-api/src/main/java/com/liferay/site/util/SitemapURLProvider.java
@@ -1,32 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.site.util;
-
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.model.LayoutSet;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.xml.Element;
-
-import org.osgi.annotation.versioning.ProviderType;
-
-/**
- * @author Eduardo García
- */
-@ProviderType
-public interface SitemapURLProvider {
-
-	public String getClassName();
-
-	public void visitLayout(
-			Element element, String layoutUuid, LayoutSet layoutSet,
-			ThemeDisplay themeDisplay)
-		throws PortalException;
-
-	public void visitLayoutSet(
-			Element element, LayoutSet layoutSet, ThemeDisplay themeDisplay)
-		throws PortalException;
-
-}
