+++ modules/apps/site/site-item-selector-web/src/main/java/com/liferay/site/item/selector/web/internal/MyGroupItemSelectorView.java
@@ -16,8 +16,8 @@ import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.site.item.selector.web.internal.renderer.MyGroupItemSelectorViewRenderer;
-import com.liferay.site.util.GroupSearchProvider;
-import com.liferay.site.util.GroupURLProvider;
+import com.liferay.site.provider.GroupSearchProvider;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.io.IOException;
