+++ modules/apps/site/site-item-selector-web/src/main/java/com/liferay/site/item/selector/web/internal/MySitesItemSelectorView.java
@@ -18,8 +18,8 @@ import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.site.item.selector.criterion.SiteItemSelectorCriterion;
 import com.liferay.site.item.selector.web.internal.renderer.MyGroupItemSelectorViewRenderer;
-import com.liferay.site.util.GroupSearchProvider;
-import com.liferay.site.util.GroupURLProvider;
+import com.liferay.site.provider.GroupSearchProvider;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.io.IOException;
