+++ modules/apps/site/site-item-selector-web/src/main/java/com/liferay/site/item/selector/web/internal/RecentSitesItemSelectorView.java
@@ -16,8 +16,8 @@ import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.site.item.selector.criterion.SiteItemSelectorCriterion;
 import com.liferay.site.item.selector.web.internal.renderer.RecentGroupItemSelectorViewRenderer;
-import com.liferay.site.util.GroupURLProvider;
-import com.liferay.site.util.RecentGroupManager;
+import com.liferay.site.manager.RecentGroupManager;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.io.IOException;
