+++ modules/apps/site/site-item-selector-web/src/main/java/com/liferay/site/item/selector/web/internal/frontend/taglib/clay/servlet/taglib/SiteVerticalCard.java
@@ -19,7 +19,7 @@ import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.site.constants.SiteWebKeys;
-import com.liferay.site.util.GroupURLProvider;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.util.Collections;
 import java.util.List;
