+++ modules/apps/site/site-item-selector-web/src/main/java/com/liferay/site/item/selector/web/internal/renderer/MyGroupItemSelectorViewRenderer.java
@@ -9,8 +9,8 @@ import com.liferay.item.selector.criteria.group.criterion.GroupItemSelectorCrite
 import com.liferay.site.constants.SiteWebKeys;
 import com.liferay.site.item.selector.web.internal.constants.SitesItemSelectorWebKeys;
 import com.liferay.site.item.selector.web.internal.display.context.MySitesItemSelectorViewDisplayContext;
-import com.liferay.site.util.GroupSearchProvider;
-import com.liferay.site.util.GroupURLProvider;
+import com.liferay.site.provider.GroupSearchProvider;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.io.IOException;
