+++ modules/apps/site/site-item-selector-web/src/main/java/com/liferay/site/item/selector/web/internal/renderer/SiteItemSelectorViewRendererImpl.java
@@ -10,7 +10,7 @@ import com.liferay.site.constants.SiteWebKeys;
 import com.liferay.site.item.selector.display.context.SitesItemSelectorViewDisplayContext;
 import com.liferay.site.item.selector.renderer.SiteItemSelectorViewRenderer;
 import com.liferay.site.item.selector.web.internal.constants.SitesItemSelectorWebKeys;
-import com.liferay.site.util.GroupURLProvider;
+import com.liferay.site.provider.GroupURLProvider;
 
 import java.io.IOException;
